+++ tests/testthat/helper-utils.R
@@ -16,6 +16,44 @@ expect_equivalent <- function(x, y) {
   expect_equal(x, y, ignore_attr = TRUE)
 }
 
+# code taken from languageserver/R/utils.R
+is_rmarkdown <- function(uri) {
+  filename <- path_from_uri(uri)
+  endsWith(tolower(filename), ".rmd") || endsWith(tolower(filename), ".rmarkdown")
+}
+
+# code taken from languageserver/R/utils.R
+path_from_uri <- function(uri) {
+  if (length(uri) == 0) {
+    return(character())
+  }
+
+  if (startsWith(uri, "file:///")) {
+    start_char <- if (.Platform$OS.type == "windows") 9 else 8
+    path <- substr(uri, start_char, nchar(uri))
+  } else if (startsWith(uri, "vscode-notebook-cell:")) {
+    # Windows: vscode-notebook-cell:/c:/Users/Username/Documents/Notebooks/MyNotebook.ipynb#MyCellId
+    # Unix: vscode-notebook-cell:/home/username/Documents/Notebooks/MyNotebook.ipynb#MyCellId
+    # WSL: vscode-notebook-cell://wsl+ubuntu-20.04/home/username/Documents/Notebooks/MyNotebook.ipynb#MyCellId
+    if (.Platform$OS.type == "windows") {
+      path <- sub("^vscode-notebook-cell:/(.+)#.*$", "\\1", uri)
+    } else {
+      path <- sub("^vscode-notebook-cell:(.+)#.*$", "\\1", uri)
+      if (startsWith(path, "//")) {
+        path <- sub("^//[^/]+(/.+)$", "\\1", path)
+      }
+    }
+  } else {
+    return("")
+  }
+
+  # URLdecode gives unknown encoding, we need to mark them as UTF-8
+  path <- utils::URLdecode(path)
+  Encoding(path) <- "UTF-8"
+  path
+}
+
+# code largely taken from languageserver/tests/testthat/helper-utils.R
 language_client <- function(working_dir = getwd(), diagnostics = FALSE, capabilities = NULL) {
   withr::local_dir(working_dir)
   withr::local_file(".Rprofile", {
@@ -45,14 +83,17 @@ language_client <- function(working_dir = getwd(), diagnostics = FALSE, capabili
 
   if (nzchar(Sys.getenv("R_LANGSVR_LOG"))) {
     script <- sprintf(
-      "languageserver::run(debug = '%s')",
+      # "languageserver::run(debug = '%s')",
+      "options(languageserver.formatting_style = NULL); languageserver::run(debug = '%s')",
       normalizePath(Sys.getenv("R_LANGSVR_LOG"), "/", mustWork = FALSE))
   } else {
-    script <- "languageserver::run()"
+    # script <- "languageserver::run()"
+    script <- "options(languageserver.formatting_style = NULL); languageserver::run()"
   }
 
   client <- languageserver:::LanguageClient$new(
-    file.path(R.home("bin"), "R"), c("--slave", "-e", script))
+    # file.path(R.home("bin"), "R"), c("--slave", "-e", script))
+    file.path(R.home("bin"), "R"), c("--no-echo", "-e", script))
 
   client$notification_handlers <- list(
     `textDocument/publishDiagnostics` = function(self, params) {
@@ -73,15 +114,23 @@ language_client <- function(working_dir = getwd(), diagnostics = FALSE, capabili
   withr::defer_parent({
     # it is sometimes necessary to shutdown the server probably
     # we skip this for other times for speed
-    if (Sys.getenv("R_LANGSVR_TEST_FAST", "YES") == "NO") {
-      client %>% respond("shutdown", NULL, retry = FALSE)
-      client$process$wait(10 * 1000)  # 10 sec
-      if (client$process$is_alive()) {
-        cat("server did not shutdown peacefully\n")
-        client$process$kill_tree()
+    # if (Sys.getenv("R_LANGSVR_TEST_FAST", "YES") == "NO") {
+    #   client %>% respond("shutdown", NULL, retry = FALSE)
+    #   client$process$wait(10 * 1000)  # 10 sec
+    #   if (client$process$is_alive()) {
+    #     cat("server did not shutdown peacefully\n")
+    #     client$process$kill_tree()
+    #   }
+    # } else {
+    #   client$process$kill_tree()
+    client %>% respond("shutdown", NULL, retry = FALSE)
+    if (client$process$is_alive()) {
+      if (identical(Sys.getenv("R_COVR"), "true")) {
+        client$process$wait()
+      } else {
+        client$process$wait(1000)
+        client$process$kill()
       }
-    } else {
-      client$process$kill_tree()
     }
   })
   client
@@ -116,6 +165,7 @@ did_open <- function(client, path, uri = languageserver:::path_to_uri(path), tex
       )
     )
   )
+  Sys.sleep(0.5)
   invisible(client)
 }
 
@@ -188,7 +238,7 @@ respond <- function(client, method, params, timeout, allow_error = FALSE,
       return(NULL)
     }
     Sys.sleep(0.2)
-    return(Recall(client, method, params, remaining, allow_error, retry, retry_when))
+return(Recall(client, method, params, remaining, allow_error, retry, retry_when))
   }
   return(result)
 }
@@ -485,6 +535,31 @@ respond_code_action <- function(client, path, start_pos, end_pos, ..., uri = lan
   )
 }
 
+respond_semantic_tokens_full <- function(client, path, ..., uri = path_to_uri(path)) {
+  respond(
+    client,
+    "textDocument/semanticTokens/full",
+    list(
+      textDocument = list(uri = uri)
+    ),
+    ...
+  )
+}
+respond_semantic_tokens_range <- function(client, path, start_pos, end_pos, ..., uri = path_to_uri(path)) {
+  respond(
+    client,
+    "textDocument/semanticTokens/range",
+    list(
+      textDocument = list(uri = uri),
+      range = range(
+        start = position(start_pos[1], start_pos[2]),
+        end = position(end_pos[1], end_pos[2])
+      )
+    ),
+    ...
+  )
+}
+
 wait_for <- function(client, method, timeout = 30) {
   storage <- new.env(parent = .GlobalEnv)
   start_time <- Sys.time()
@@ -510,5 +585,33 @@ wait_for <- function(client, method, timeout = 30) {
     remaining <- (start_time + timeout) - Sys.time()
   }
   NULL
+}
+
+respond_prepare_type_hierarchy <- function(client, path, pos, ..., uri = path_to_uri(path)) {
+  respond(
+    client,
+    "textDocument/prepareTypeHierarchy",
+    list(
+      textDocument = list(uri = uri),
+      position = list(line = pos[1], character = pos[2])
+    ),
+    ...
+  )
+}
+respond_type_hierarchy_supertypes <- function(client, item, ...) {
+  respond(
+    client,
+    "typeHierarchy/supertypes",
+    list(item = item),
+    ...
+  )
+}
+respond_type_hierarchy_subtypes <- function(client, item, ...) {
+  respond(
+    client,
+    "typeHierarchy/subtypes",
+    list(item = item),
+    ...
+  )
 }
 # nolint end