+++ tests/testthat/test-lsp_aaa_regression.R
@@ -12,8 +12,8 @@ test_that("Completion of attached package functions works", {
     ),
     temp_file
   )
-
-  client %>% did_save(temp_file)
+# browser()
+  client %>% did_open(temp_file)
 
   result <- client %>% respond_completion(
     temp_file,